requirements.txt
pytest>=8.0
allure-pytest>=2.15.0
selenium>=4.39.0
webdriver-manager>=4.0.0
python-dotenv==1.2.2
